Django_Application/requirements.txt
django-crispy-forms==1.7.2
django-extensions==2.0.7
django-filter==1.1.0
django-json-response==1.1.5
django-mysql==2.2.2
djangorestframework==3.8.2
djangorestframework-files==1.1.0
mysqlclient==1.3.12
pip==10.0.1
psutil==5.4.6
pytz==2018.4
setuptools==39.2.0
six==1.11.0
typing==3.6.4
wheel==0.31.1 
djangorestframework-jwt==1.11.0 
